StochasticInterpolants

Documentation for StochasticInterpolants.

StochasticInterpolants.StochasticInterpolants

StochasticInterpolants.AttnParsConvNextUNet

StochasticInterpolants.Autoencoder

StochasticInterpolants.ConditionalDiffusionTransformer

StochasticInterpolants.ConditionalTransformerUNet

StochasticInterpolants.ConvNextUNetNoPars

StochasticInterpolants.ConvNextUNetWithPars

StochasticInterpolants.DataDependentCouplingStochasticInterpolant

StochasticInterpolants.DataDependentCouplingStochasticInterpolant

StochasticInterpolants.DataDependentCouplingStochasticInterpolant

StochasticInterpolants.DenoisingDiffusionProbabilisticModel

StochasticInterpolants.DiffusionTransformer

StochasticInterpolants.DiffusionTransformerBlock

StochasticInterpolants.EncoderFollmerStochasticInterpolant

StochasticInterpolants.EncoderFollmerStochasticInterpolant

StochasticInterpolants.EntropicActionMatching

StochasticInterpolants.EntropicActionMatching

StochasticInterpolants.FollmerStochasticInterpolant

StochasticInterpolants.FollmerStochasticInterpolant

StochasticInterpolants.LatentFollmerStochasticInterpolant

StochasticInterpolants.LatentFollmerStochasticInterpolant

StochasticInterpolants.NoiseScheduling

StochasticInterpolants.NormalizePars

StochasticInterpolants.PhysicsInformedStochasticInterpolant

StochasticInterpolants.PhysicsInformedStochasticInterpolant

StochasticInterpolants.PositionEncoding

StochasticInterpolants.ScoreMatchingLangevinDynamics

StochasticInterpolants.StandardizeData

StochasticInterpolants.StochasticInterpolantModel

StochasticInterpolants.StochasticInterpolantModel

StochasticInterpolants.ViPosEmbedding

StochasticInterpolants.CheckpointManager

StochasticInterpolants.Interpolant

StochasticInterpolants.LinearMultiHeadSelfAttention

StochasticInterpolants.MultiHeadSelfAttention

StochasticInterpolants.VisionTransformerEncoder

StochasticInterpolants._fast_chunk

StochasticInterpolants._flatten_spatial

StochasticInterpolants.conv_next_block

StochasticInterpolants.conv_next_block_no_pars

StochasticInterpolants.create_checkpoint_directory

StochasticInterpolants.euler_maruyama_sampler

StochasticInterpolants.forward_diffusion_sample

StochasticInterpolants.get_beta_schedule

StochasticInterpolants.get_encoder_forecasting_loss

StochasticInterpolants.get_forecasting_from_gaussian_loss

StochasticInterpolants.get_forecasting_loss

StochasticInterpolants.get_forecasting_loss

StochasticInterpolants.get_loss

StochasticInterpolants.get_loss

StochasticInterpolants.get_loss

StochasticInterpolants.get_physics_forecasting_loss

StochasticInterpolants.get_t_pars_embedding

StochasticInterpolants.load_model_config

StochasticInterpolants.load_model_weights

StochasticInterpolants.multiple_conv_next_blocks

StochasticInterpolants.multiple_conv_next_blocks_no_pars

StochasticInterpolants.parameter_diffusion_transformer_block

StochasticInterpolants.patchify

StochasticInterpolants.sample_timestep

StochasticInterpolants.save_checkpoint

StochasticInterpolants.sinusoidal_embedding

StochasticInterpolants.train_diffusion_model

StochasticInterpolants.train_stochastic_interpolant

StochasticInterpolants.train_stochastic_interpolant

StochasticInterpolants.train_stochastic_interpolant_for_closure

StochasticInterpolants.unpatchify

StochasticInterpolants.StochasticInterpolants — Module

StochasticInterpolants

Implementation of the Stochastic Interpolants method for generative modeling.

StochasticInterpolants.AttnParsConvNextUNet — Type

AttnParsConvNextUNet( imagesize::Tuple{Int, Int}, channels::Vector{Int} = [32, 64, 96, 128], blockdepth::Int = 2, minfreq::Float32 = 1.0f0, maxfreq::Float32 = 1000.0f0, embedding_dims::Int = 32 )

Create a conditional U-Net model with the given image size, number of channels, block depth, frequency range, and embedding dimensions.

StochasticInterpolants.Autoencoder — Type

Autoencoder(in_channels::Int = 3, channels=[32, 64, 128], padding="periodic")

Create an autoencoder model with the given number of input channels, output channels, and padding type.

StochasticInterpolants.ConditionalDiffusionTransformer — Type

ConditionalDiffusionTransformer( imsize::Dims{2}=(256, 256), inchannels::Int=3, patchsize::Dims{2}=(16, 16), embedplanes::Int=768, depth::Int=6, numberheads=16, mlpratio=4.0f0, dropoutrate=0.1f0, embeddingdropoutrate=0.1f0, pool::Symbol=:class, num_classes::Int=1000, kwargs... )

Create a DiffusionTransformer Transformer model with the given image size, number of input channels, patch size, embedding planes, depth, number of heads, MLP ratio, dropout rate, embedding dropout rate, pooling method, number of classes, and additional keyword arguments.

Based on https://github.com/LuxDL/Boltz.jl/blob/v0.3.9/src/vision/vit.jl#L48-L61 The architecture is based on https://arxiv.org/abs/2212.09748

StochasticInterpolants.ConditionalTransformerUNet — Type

ConditionalTransformerUNet(
    image_size::Tuple{Int, Int},
    channels::Vector{Int} = [32, 64, 96, 128],
    block_depth::Int = 2,
    min_freq::Float32 = 1.0f0,
    max_freq::Float32 = 1000.0f0,
    embedding_dims::Int = 32
)

Create a conditional U-Net model with the given image size, number of channels, block depth, frequency range, and embedding dimensions.

StochasticInterpolants.ConvNextUNetNoPars — Type

ConvNextUNetNoPars( imagesize::Tuple{Int, Int}, channels::Vector{Int} = [32, 64, 96, 128], blockdepth::Int = 2, minfreq::Float32 = 1.0f0, maxfreq::Float32 = 1000.0f0, )

Create a conditional U-Net model with the given image size, number of channels, block depth, frequency range, and embedding dimensions.

StochasticInterpolants.ConvNextUNetWithPars — Type

ConvNextUNetWithPars( imagesize::Tuple{Int, Int}, channels::Vector{Int} = [32, 64, 96, 128], blockdepth::Int = 2, minfreq::Float32 = 1.0f0, maxfreq::Float32 = 1000.0f0, embedding_dims::Int = 32 )

Create a conditional U-Net model with the given image size, number of channels, block depth, frequency range, and embedding dimensions.

StochasticInterpolants.DataDependentCouplingStochasticInterpolant — Type

DataDependentCouplingStochasticInterpolant(
    velocity::Lux.AbstractExplicitLayer
    score::Lux.AbstractExplicitLayer

)

A container layer for the Stochastic Interpolant model

StochasticInterpolants.DataDependentCouplingStochasticInterpolant — Method

DataDependentCouplingStochasticInterpolant( velocity::Lux.AbstractExplicitLayer; interpolant=Interpolant(), gamma=Gamma(), diffusioncoefficient=DiffusionCoefficient(), diffusionmultiplier=0.1f0, dev=gpu_device() interpolant::Interpolant )

Constructs a Stochastic Interpolant model

StochasticInterpolants.DataDependentCouplingStochasticInterpolant — Method

DataDependentCouplingStochasticInterpolant(
    velocity::Lux.AbstractExplicitLayer; 
    interpolant=Interpolant(),
    diffusion_multiplier=0.1f0,
    dev=gpu_device() 
)

Constructs a Stochastic Interpolant model

StochasticInterpolants.DenoisingDiffusionProbabilisticModel — Type

DenoisingDiffusionProbabilisticModel(
    image_size::Tuple{Int, Int}; 
    in_channels::Int = 3,
    channels=[32, 64, 96, 128], 
    block_depth=2,
    min_freq=1.0f0, 
    max_freq=1000.0f0, 
    embedding_dims=32,
    timesteps::Int=100,
    init_beta::AbstractFloat=0.0001, 
    final_beta::AbstractFloat=0.02,
    type::String="linear"
)

Creates a Denoising Diffusion Probabilistic Model (DDPM) with a UNet architecture.

StochasticInterpolants.DiffusionTransformer — Type

DiffusionTransformer(
    imsize::Dims{2}=(256, 256),
    in_channels::Int=3,
    patch_size::Dims{2}=(16, 16),
    embed_planes::Int=768,
    depth::Int=6,
    number_heads=16,
    mlp_ratio=4.0f0,
    dropout_rate=0.1f0,
    embedding_dropout_rate=0.1f0,
    pool::Symbol=:class,
    num_classes::Int=1000,
    kwargs...
)

Create a DiffusionTransformer Transformer model with the given image size, number of input channels, patch size, embedding planes, depth, number of heads, MLP ratio, dropout rate, embedding dropout rate, pooling method, number of classes, and additional keyword arguments.

Based on https://github.com/LuxDL/Boltz.jl/blob/v0.3.9/src/vision/vit.jl#L48-L61 The architecture is based on https://arxiv.org/abs/2212.09748

StochasticInterpolants.DiffusionTransformerBlock — Type

DiffusionTransformerBlock(
    in_channels::Int, 
    out_channels::Int,
    block_depth::Int
)

Create a diffusion transformer block

StochasticInterpolants.EncoderFollmerStochasticInterpolant — Type

EncoderFollmerStochasticInterpolant(
    unet::UNet,
    sde_sample::Function,
    ode_sample::Function,
    interpolant::Function
)

A container layer for the Stochastic Interpolant model

StochasticInterpolants.EncoderFollmerStochasticInterpolant — Method

EncoderFollmerStochasticInterpolant(
    velocity::Lux.AbstractExplicitLayer; 
    interpolant=Interpolant(),
    diffusion_multiplier=0.1f0,
    dev=gpu_device() 
)

Constructs a Stochastic Interpolant model

StochasticInterpolants.EntropicActionMatching — Type

FollmerStochasticInterpolant(
)

A container layer for the Stochastic Interpolant model

StochasticInterpolants.EntropicActionMatching — Method

EntropicActionMatching(
    velocity::Lux.AbstractExplicitLayer; 
    diffusion_coefficient=DiffusionCoefficient(t -> sqrt.((3f0 .- t) .* (1f0 .- t))),
    projection=nothing,
    dev=gpu_device()
)

Constructs a Stochastic Interpolant model

StochasticInterpolants.FollmerStochasticInterpolant — Type

FollmerStochasticInterpolant(
    unet::UNet,
    sde_sample::Function,
    ode_sample::Function,
    interpolant::Function
)

A container layer for the Stochastic Interpolant model

StochasticInterpolants.FollmerStochasticInterpolant — Method

FollmerStochasticInterpolant(
    velocity::Lux.AbstractExplicitLayer; 
    interpolant=Interpolant(),
    diffusion_multiplier=0.1f0,
    dev=gpu_device() 
)

Constructs a Stochastic Interpolant model

StochasticInterpolants.LatentFollmerStochasticInterpolant — Type

LatentFollmerStochasticInterpolant(
    unet::UNet,
    sde_sample::Function,
    ode_sample::Function,
    interpolant::Function
)

A container layer for the Stochastic Interpolant model

StochasticInterpolants.LatentFollmerStochasticInterpolant — Method

LatentFollmerStochasticInterpolant(
    velocity::Lux.AbstractExplicitLayer,
    autoencoder::Lux.AbstractExplicitLayer;
    interpolant=Interpolant(),
    diffusion_multiplier=0.1f0,
    dev=gpu_device() 
)

Constructs a Stochastic Interpolant model

StochasticInterpolants.NoiseScheduling — Type

NoiseScheduling

A struct to hold the noise scheduling parameters.

This struct holds the parameters for the noise scheduling process.

Fields

betas::Vector{T}: The beta values.

alphas::Vector{T}: The alpha values.

alphas_cumprod::Vector{T}: The cumulative product of the alpha values.

alphas_cumprod_prev::Vector{T}: The cumulative product of the alpha values, shifted by one.

sqrt_recip_alphas::Vector{T}: The square root of the reciprocal of the alpha values.

sqrt_alphas_cumprod::Vector{T}: The square root of the cumulative product of the alpha values.

sqrt_one_minus_alphas_cumprod::Vector{T}: The square root of one minus the cumulative product of the alpha values.

posterior_variance::Vector{T}: The posterior variance.

StochasticInterpolants.NormalizePars — Type

NormalizePars(mean, std)

A container for the normalization and inverse normalization functions. Data is transformed to be between 0 and 1.

StochasticInterpolants.PhysicsInformedStochasticInterpolant — Type

PhysicsInformedStochasticInterpolant(
    unet::UNet,
    sde_sample::Function,
    ode_sample::Function,
    interpolant::Function
)

A container layer for the Stochastic Interpolant model

StochasticInterpolants.PhysicsInformedStochasticInterpolant — Method

PhysicsInformedStochasticInterpolant(
    velocity::Lux.AbstractExplicitLayer; 
    interpolant=Interpolant(),
    diffusion_multiplier=0.1f0,
    dev=gpu_device() 
)

Constructs a Stochastic Interpolant model

StochasticInterpolants.PositionEncoding — Type

position_encoding(dim_embedding::Int, max_length::Int=1000)

Create a position encoding for a transformer model.

Based on https://liorsinai.github.io/machine-learning/2022/05/18/transformers.html#position-encodings

StochasticInterpolants.ScoreMatchingLangevinDynamics — Type

ScoreMatchingLangevinDynamics( imagesize::Tuple{Int, Int}; inchannels::Int = 3, channels=[32, 64, 96, 128], blockdepth=2, minfreq=1.0f0, maxfreq=1000.0f0, embeddingdims=32, eps=1e-5, )

Creates a Denoising Diffusion Probabilistic Model (DDPM) with a UNet architecture.

StochasticInterpolants.StandardizeData — Type

StandardizeData(mean, std)

A container for the standardization and inverse standardization functions. Data is transform to have zero mean and unit variance.

StochasticInterpolants.StochasticInterpolantModel — Type

StochasticInterpolant(
    unet::UNet,
    sde_sample::Function,
    ode_sample::Function,
    interpolant::Function
)

A container layer for the Stochastic Interpolant model

StochasticInterpolants.StochasticInterpolantModel — Method

StochasticInterpolantModel(
    image_size::Tuple{Int, Int}; 
    in_channels::Int = 3,
    channels=[32, 64, 96, 128], 
    block_depth=2,
    min_freq=1.0f0,
    max_freq=1000.0f0,
    embedding_dims=32,
    num_steps=100,
)

Constructs a Stochastic Interpolant model

StochasticInterpolants.ViPosEmbedding — Type

ViPosEmbedding(embedding_size, number_patches; init = randn32)

Positional embedding layer used by many vision transformer-like models.

StochasticInterpolants.CheckpointManager — Method

CheckpointManager(
    test_problem::String,
    model_name::String,
    config::Dict;
    base_folder::String = "trained_models"
)

StochasticInterpolants.Interpolant — Function

Interpolant(
    alpha, 
    beta, 
    dalpha_dt, 
    dbeta_dt, 
    interpolant, 
    dinterpolant_dt
)

A struct that contains the functions for the interpolation and its derivatives.

StochasticInterpolants.LinearMultiHeadSelfAttention — Method

LinearMultiHeadSelfAttention(in_planes::Int, number_heads::Int; qkv_bias::Bool=false,
    attention_dropout_rate::T=0.0f0, projection_dropout_rate::T=0.0f0)

Linear Multi-head self-attention layer based on the paper "Efficient Attention: Attention with Linear Complexities"

StochasticInterpolants.MultiHeadSelfAttention — Method

MultiHeadSelfAttention(in_planes::Int, number_heads::Int; qkv_bias::Bool=false,
    attention_dropout_rate::T=0.0f0, projection_dropout_rate::T=0.0f0)

Multi-head self-attention layer

Arguments

planes: number of input channels

nheads: number of heads

qkv_bias: whether to use bias in the layer to get the query, key and value

attn_dropout_prob: dropout probability after the self-attention layer

proj_dropout_prob: dropout probability after the projection layer

StochasticInterpolants.VisionTransformerEncoder — Method

VisionTransformerEncoder(in_planes, depth, number_heads; mlp_ratio = 4.0f0,
    dropout = 0.0f0)

Transformer as used in the base ViT architecture.

Arguments

in_planes: number of input channels

depth: number of attention blocks

number_heads: number of attention heads

Keyword Arguments

mlp_ratio: ratio of MLP layers to the number of input channels

dropout_rate: dropout rate

References

[1] Dosovitskiy, Alexey, et al. "An image is worth 16x16 words: Transformers for image recognition at scale." arXiv preprint arXiv:2010.11929 (2020).

StochasticInterpolants._fast_chunk — Method

_fast_chunk(x::AbstractArray, ::Val{n}, ::Val{dim})

Type-stable and faster version of MLUtils.chunk.

StochasticInterpolants._flatten_spatial — Method

_flatten_spatial(x::AbstractArray{T, 4})

Flattens the first 2 dimensions of x, and permutes the remaining dimensions to (2, 1, 3)

StochasticInterpolants.conv_next_block — Method

conv_next_block(
    in_channels::Int, 
    out_channels::Int,
    multiplier::Int = 1,
    pars_embed_dim::Int = 1,
    imsize::Tuple{Int, Int} = (64, 128)
)

Create a conv next block with the given number of input and output channels. The block consists of two convolutional layers with kernel size kernel_size. The first layer has the same number of input and output channels, while the second layer has the same number of output channels as the block. The block also includes batch normalization and a skip connection.

https://arxiv.org/abs/2201.03545

Based on https://github.com/tum-pbs/autoreg-pde-diffusion/blob/b9b33913b99ede88d9452c5ab470c5d7f5da5c56/src/turbpred/modeldiffusionblocks.py#L60

StochasticInterpolants.conv_next_block_no_pars — Method

conv_next_block_no_pars(
    in_channels::Int, 
    out_channels::Int,
    multiplier::Int = 1,
    pars_embed_dim::Int = 1,
    imsize::Tuple{Int, Int} = (64, 128)
)

Create a conv next block with the given number of input and output channels. The block consists of two convolutional layers with kernel size kernel_size. The first layer has the same number of input and output channels, while the second layer has the same number of output channels as the block. The block also includes batch normalization and a skip connection.

https://arxiv.org/abs/2201.03545

Based on https://github.com/tum-pbs/autoreg-pde-diffusion/blob/b9b33913b99ede88d9452c5ab470c5d7f5da5c56/src/turbpred/modeldiffusionblocks.py#L60

StochasticInterpolants.create_checkpoint_directory — Method

create_checkpoint_directory(
    test_problem::String,
    model_name::String;
    base_folder::String = "trained_models"
)

StochasticInterpolants.euler_maruyama_sampler — Function

Euler_Maruyama_sampler(
    model::ScoreMatchingLangevinDynamics,
    marginal_probability_std::Function,
    ps::NamedTuple,
    st::NamedTuple,
    rng::AbstractRNG,
    num_samples::Int,
    num_steps::Int,
    eps::Float32,
    dev=gpu_device()
)

Samples from the SMLD model using the Euler-Maruyama method

StochasticInterpolants.forward_diffusion_sample — Function

forward_diffusion_sample(
    x_0::AbstractArray, 
    t::AbstractArray, 
    rng::AbstractRNG, 
    noise_scheduling::NoiseScheduling, 
    dev=gpu_device()
)

Samples the forward diffusion process.

StochasticInterpolants.get_beta_schedule — Method

getbetaschedule

Returns a schedule of beta values for the diffusion process.

Based on https://colab.research.google.com/drive/1sjy9odlSSy0RBVgMTgP7s99NXsqglsUL?usp=sharing#scrollTo=qWw50ui9IZ5q

StochasticInterpolants.get_encoder_forecasting_loss — Function

get_forecasting_loss(
    x_0::AbstractArray, 
    x_1::AbstractArray,
    pars::AbstractArray,
    velocity::Lux.AbstractExplicitLayer,
    encoder::Lux.AbstractExplicitLayer,
    interpolant::Function, 
    gamma::Function,
    ps::NamedTuple, 
    st::NamedTuple,
    rng::AbstractRNG,
    dev=gpu_device()
)

Computes the loss for the stochastic interpolant Model.

StochasticInterpolants.get_forecasting_from_gaussian_loss — Function

get_forecasting_from_gaussian_loss(
    x_1::AbstractArray,
    x_condition::AbstractArray,
    pars::AbstractArray,
    velocity::Lux.AbstractExplicitLayer,
    interpolant::NamedTuple, 
    ps::NamedTuple, 
    st::NamedTuple,
    rng::AbstractRNG,
    dev=gpu_device()
)

Computes the loss for the stochastic interpolant Model.

StochasticInterpolants.get_forecasting_loss — Function

get_forecasting_loss(
    x_0::AbstractArray, 
    x_1::AbstractArray,
    pars::AbstractArray,
    velocity::UNet,
    interpolant::Function, 
    gamma::Function,
    ps::NamedTuple, 
    st::NamedTuple,
    rng::AbstractRNG,
    dev=gpu_device()
)

Computes the loss for the stochastic interpolant Model.

StochasticInterpolants.get_forecasting_loss — Function

get_forecasting_loss(
    x_0::AbstractArray, 
    x_1::AbstractArray,
    pars::AbstractArray,
    velocity::UNet,
    interpolant::Function, 
    gamma::Function,
    ps::NamedTuple, 
    st::NamedTuple,
    rng::AbstractRNG,
    dev=gpu_device()
)

Computes the loss for the stochastic interpolant Model.

StochasticInterpolants.get_loss — Function

get_loss(
    x_0::AbstractArray, 
    x_1::AbstractArray,
    t::AbstractArray, 
    velocity::UNet,
    interpolant::Function, 
    gamma::Function,
    ps::NamedTuple, 
    st::NamedTuple,
    rng::AbstractRNG,
    dev=gpu_device()
)

Computes the loss for the stochastic interpolant Model.

StochasticInterpolants.get_loss — Function

get_loss(
    x_0::AbstractArray, 
    t::AbstractArray, 
    model::ScoreMatchingLangevinDynamics, 
    marginal_prob_std, 
    eps=1e-5,
    dev=gpu_device()
)

Computes the loss of the SMLD model

StochasticInterpolants.get_loss — Function

get_loss(
    x_0::AbstractArray, 
    x_1::AbstractArray,
    t::AbstractArray, 
    velocity::UNet,
    score::UNet,
    interpolant::Function, 
    gamma::Function,
    ps::NamedTuple, 
    st::NamedTuple,
    rng::AbstractRNG,
    dev=gpu_device()
)

Computes the loss for the stochastic interpolant Model.

StochasticInterpolants.get_physics_forecasting_loss — Function

get_physics_forecasting_loss(
    x_0::AbstractArray, 
    x_1::AbstractArray,
    pars::AbstractArray,
    velocity::UNet,
    interpolant::Function, 
    gamma::Function,
    ps::NamedTuple, 
    st::NamedTuple,
    rng::AbstractRNG,
    dev=gpu_device()
)

Computes the loss for the stochastic interpolant Model.

StochasticInterpolants.get_t_pars_embedding — Function

get_t_pars_embedding(
    pars_dim,
    with_time,
    embedding_dim,
    min_freq=1.0f0, 
    max_freq=1000.0f0, 
)

Get the time and parameter embedding layer.

StochasticInterpolants.load_model_config — Method

load_model_config(path::String)

StochasticInterpolants.load_model_weights — Method

load_model_weights(path::String)

StochasticInterpolants.multiple_conv_next_blocks — Method

multiple_conv_next_blocks(
    in_channels::Int, 
    out_channels::Int,
    multiplier::Int = 1,
    embedding_dims::Int = 1,
    imsize::Tuple{Int, Int} = (64, 128)
)

Create a chain of two conv next blocks with the given number of input and output channels. The first block has the same number of input and output

StochasticInterpolants.multiple_conv_next_blocks_no_pars — Method

multiple_conv_next_blocks_no_pars(
    in_channels::Int, 
    out_channels::Int,
    multiplier::Int = 1,
    padding="constant"
)

Create a chain of two conv next blocks with the given number of input and output channels. The first block has the same number of input and output

StochasticInterpolants.parameter_diffusion_transformer_block — Method

parameter_diffusion_transformer_block(
    in_channels::Int,
    out_channels::Int,
    pars_dim::Int,
    embed_dim::Int,
    number_heads::Int,
    mlp_ratio::Int,
    imsize::Tuple{Int, Int},
    patch_size::Int,
    number_patches::Int
)

Create a parameter diffusion transformer block with the given number of input and output channels, parameter dimensions, embedding dimensions, number of heads, MLP ratio, image size, patch size, and number of patches.

StochasticInterpolants.patchify — Method

patchify(
    imsize::Tuple{Int, Int}=(64, 64),
    in_channels::Int=3,
    patch_size::Tuple{Int, Int}=(8, 8),
    embed_planes::Int=128,
    norm_layer=Returns(Lux.NoOpLayer()),
    flatten=true
)

Create a patch embedding layer with the given image size, number of input channels, patch size, embedding planes, normalization layer, and flatten flag.

Based on https://github.com/LuxDL/Boltz.jl/blob/v0.3.9/src/vision/vit.jl#L48-L61

StochasticInterpolants.sample_timestep — Method

sample_timestep(
    x::AbstractArray, 
    t::AbstractArray;
    model,
    noise_scheduling::NoiseScheduling,
    ps::NamedTuple,
    st::NamedTuple,
    rng::AbstractRNG,
    dev=gpu_device()
)

Calls the model to predict the noise in the image and returns the denoised image. Applies noise to this image, if we are not in the last step yet.

StochasticInterpolants.save_checkpoint — Method

save_checkpoint(ps, st, opt_st, output_dir, epoch)

StochasticInterpolants.sinusoidal_embedding — Function

sinusoidal_embedding(
    x::AbstractArray{AbstractFloat, 4},
    min_freq::AbstractFloat,
    max_freq::AbstractFloat,
    embedding_dims::Int
)

Embed the noise variances to a sinusoidal embedding with the given frequency range and embedding dimensions.

Based on https://yng87.page/en/blog/2022/lux-ddim/.

StochasticInterpolants.train_diffusion_model — Function

train_diffusion_model(
    model,
    ps::NamedTuple,
    st::NamedTuple,
    opt_state::NamedTuple,
    trainset::AbstractArray,
    num_epochs::Int,
    batch_size::Int,
    num_samples::Int,
    rng::AbstractRNG,
    dev=gpu_device()
)

Train the Diffusion model.

StochasticInterpolants.train_stochastic_interpolant — Function

train_stochastic_interpolant(
    model::StochasticInterpolantModel,
    ps::NamedTuple,
    st::NamedTuple,
    opt_state::NamedTuple,
    trainset::AbstractArray,
    num_epochs::Int,
    batch_size::Int,
    num_samples::Int,
    rng::AbstractRNG,
    trainset_init_distribution::AbstractArray=nothing,
    init_and_target_are_correlated=false,
    train_time_stepping=false,
    dev=gpu_device()
)

Train the Stochastic Interpolant model.

StochasticInterpolants.train_stochastic_interpolant — Method

train_stochastic_interpolant(
    model,
    ps::NamedTuple,
    st::NamedTuple,
    opt_state::NamedTuple,
    trainset::AbstractArray,
    num_epochs::Int,
    batch_size::Int,
    num_samples::Int,
    rng::AbstractRNG,
    trainset_init_distribution::AbstractArray=nothing,
    init_and_target_are_correlated=false,
    train_time_stepping=false,
    dev=gpu_device()
)

Train the Stochastic Interpolant model.

StochasticInterpolants.train_stochastic_interpolant_for_closure — Method

train_stochastic_interpolant_for_closure(
    model,
    ps::NamedTuple,
    st::NamedTuple,
    opt_state::NamedTuple,
    trainset::AbstractArray,
    num_epochs::Int,
    batch_size::Int,
    num_samples::Int,
    rng::AbstractRNG,
    trainset_init_distribution::AbstractArray=nothing,
    init_and_target_are_correlated=false,
    train_time_stepping=false,
    dev=gpu_device()
)

Train the Stochastic Interpolant model.

StochasticInterpolants.unpatchify — Method

unpatchify(x, patch_size, out_channels)

Unpatchify the input tensor x with the given patch size and number of output channels.